# fix bug in "proxy to graphite upon bad request"